Project overall revamp
