+++ routes/eFrederick/api/v1/index.js
@@ -1,0 +1,9 @@
+const express = require('express');
+
+const router = express.Router();
+
+const keys = require('./keys');
+
+router.use('/keys', keys);
+
+module.exports = router;
