+++ routes/eFrederick/index.js
@@ -4,6 +4,6 @@ const router = express.Router();
 
 const apiV1 = require('./api/v1');
 
-router.use('/v1', apiV1)
+router.use('/api/v1', apiV1);
 
 module.exports = router;
