+++ routes/namah/api/v1/banners.js
@@ -1,8 +1,8 @@
 const express = require('express');
 
 //Models
-const getConnection = require('../../../models/createPool');
-const getQuery = require('../../../models/createQuery');
+const getConnection = require('../../../../models/createPool');
+const getQuery = require('../../../../models/createQuery');
 
 const router = express.Router();
